requirements.txt
alembic==1.0.0
bcrypt==3.1.4
blinker==1.4
cffi==1.11.5
click==6.7
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.2.1
Flask-Moment==0.6.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
Pillow==5.2.0
pycparser==2.18
PyJWT==1.6.4
python-dateutil==2.7.3
python-editor==1.0.3
six==1.11.0
SQLAlchemy==1.2.10
Werkzeug==0.14.1
WTForms==2.2.1
